Preliminary commit for min const generics
