+++ ff/src/fields/mod.rs
@@ -21,8 +21,6 @@ pub use ark_ff_macros;
 use num_traits::{One, Zero};
 use zeroize::Zeroize;
 
-#[macro_use]
-pub mod macros;
 pub mod utils;
 
 #[macro_use]
@@ -525,17 +523,6 @@ impl<Slice: AsRef<[u64]>> Iterator for BitIteratorLE<Slice> {
         }
     }
 }
-
-use crate::biginteger::{
-    BigInteger256, BigInteger320, BigInteger384, BigInteger64, BigInteger768, BigInteger832,
-};
-
-impl_field_bigint_conv!(Fp64, BigInteger64, Fp64Parameters);
-impl_field_bigint_conv!(Fp256, BigInteger256, Fp256Parameters);
-impl_field_bigint_conv!(Fp320, BigInteger320, Fp320Parameters);
-impl_field_bigint_conv!(Fp384, BigInteger384, Fp384Parameters);
-impl_field_bigint_conv!(Fp768, BigInteger768, Fp768Parameters);
-impl_field_bigint_conv!(Fp832, BigInteger832, Fp832Parameters);
 
 // Given a vector of field elements {v_i}, compute the vector {v_i^(-1)}
 pub fn batch_inversion<F: Field>(v: &mut [F]) {
